requirements.txt
python-binance
urllib3<2